extracted ts-compiler in its own module
