+++ packages/runtime/lib/compile.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { tsCompiler } = require('@platformatic/service')
+const tsCompiler = require('@platformatic/ts-compiler')
 const { loadConfig } = require('./load-config')
 const { dirname } = require('node:path')
 
@@ -42,11 +42,27 @@ async function compile (argv, logger) {
         watch: false
       }, false)
 
-      const serviceWasCompiled = await tsCompiler.compile(service.path, configManager.current, childLogger, compileOptions)
+      const serviceWasCompiled = await tsCompiler.compile({
+        ...compileOptions,
+        cwd: service.path,
+        // TODO(mcollina): tsConfig and flags are @platformatic/service specific.
+        // we must generalize them
+        tsConfig: configManager.current.plugins?.typescript?.tsConfig,
+        flags: configManager.current.plugins?.typescript?.flags,
+        logger: childLogger
+      })
       compiled ||= serviceWasCompiled
     }
   } else {
-    compiled = await tsCompiler.compile(dirname(configManager.fullPath), configManager.current, logger, compileOptions)
+    compiled = await tsCompiler.compile({
+      ...compileOptions,
+      cwd: dirname(configManager.fullPath),
+      // TODO(mcollina): tsConfig and flags are @platformatic/service specific.
+      // we must generalize them
+      tsConfig: configManager.current.plugins?.typescript?.tsConfig,
+      flags: configManager.current.plugins?.typescript?.flags,
+      logger
+    })
   }
 
   return compiled
