+++ packages/service/lib/compile.js
@@ -1,107 +1,16 @@
 'use strict'
 
-const { resolve, join, dirname } = require('path')
+const { compile } = require('@platformatic/ts-compiler')
+const { loadConfig } = require('@platformatic/config')
 const pino = require('pino')
 const pretty = require('pino-pretty')
-const { loadConfig } = require('@platformatic/config')
-const { isFileAccessible } = require('./utils.js')
-const { readFile, rm } = require('fs/promises')
-
-async function getTSCExecutablePath (cwd) {
-  const typescriptPath = require.resolve('typescript')
-  const typescriptPathCWD = require.resolve('typescript', { paths: [process.cwd()] })
-
-  const tscLocalPath = join(typescriptPath, '..', '..', 'bin', 'tsc')
-  const tscGlobalPath = join(typescriptPathCWD, '..', '..', 'bin', 'tsc')
-
-  const [tscLocalExists, tscGlobalExists] = await Promise.all([
-    isFileAccessible(tscLocalPath),
-    isFileAccessible(tscGlobalPath)
-  ])
-
-  /* c8 ignore next 7 */
-  if (tscLocalExists) {
-    return tscLocalPath
-  }
-
-  if (tscGlobalExists) {
-    return tscGlobalPath
-  }
-}
-
-async function setup (cwd, config, logger) {
-  if (!logger) {
-    logger = pino(
-      pretty({
-        translateTime: 'SYS:HH:MM:ss',
-        ignore: 'hostname,pid'
-      })
-    )
-
-    if (config?.server.logger) {
-      logger.level = config.server.logger.level
-    }
-  }
-
-  const { execa } = await import('execa')
-
-  const tscExecutablePath = await getTSCExecutablePath(cwd)
-  /* c8 ignore next 4 */
-  if (tscExecutablePath === undefined) {
-    const msg = 'The tsc executable was not found.'
-    logger.warn(msg)
-  }
-
-  const tsConfigPath = config?.plugins?.typescript?.tsConfig || resolve(cwd, 'tsconfig.json')
-  const tsConfigExists = await isFileAccessible(tsConfigPath)
-
-  if (!tsConfigExists) {
-    const msg = 'No typescript configuration file was found, skipping compilation.'
-    logger.info(msg)
-  }
-
-  return { execa, logger, tscExecutablePath, tsConfigPath, tsConfigExists }
-}
-
-async function compile (cwd, config, originalLogger, options) {
-  const { execa, logger, tscExecutablePath, tsConfigPath, tsConfigExists } = await setup(cwd, config, originalLogger)
-  /* c8 ignore next 3 */
-  if (!tscExecutablePath || !tsConfigExists) {
-    return false
-  }
-
-  try {
-    const tsFlags = config?.plugins?.typescript?.flags || ['--project', tsConfigPath, '--rootDir', '.']
-    const env = {
-      ...process.env
-    }
-    delete env.NODE_V8_COVERAGE
-    // somehow c8 does not pick up these lines even if there is a specific test
-    /* c8 ignore start */
-    if (options && options.clean) {
-      // delete outdir directory
-      const tsConfigContents = JSON.parse(await readFile(tsConfigPath, 'utf8'))
-      const outDir = tsConfigContents.compilerOptions.outDir
-      if (outDir) {
-        const outDirFullPath = join(dirname(tsConfigPath), outDir)
-        originalLogger.info(`Removing build directory ${outDirFullPath}`)
-        await rm(outDirFullPath, { recursive: true })
-      }
-    }
-    /* c8 ignore stop */
-    await execa(tscExecutablePath, tsFlags, { cwd, env })
-    logger.info('Typescript compilation completed successfully.')
-    return true
-  } catch (error) {
-    logger.error(error.message)
-    return false
-  }
-}
+const { dirname } = require('path')
 
 function buildCompileCmd (app) {
   return async function compileCmd (_args) {
     let fullPath = null
     let config = null
+
     try {
       const { configManager } = await loadConfig({}, _args, app, {
         watch: false
@@ -114,9 +23,7 @@ function buildCompileCmd (app) {
       console.error(err)
       process.exit(1)
     }
-    const compileOptions = {
-      clean: _args.includes('--clean')
-    }
+
     const logger = pino(
       pretty({
         translateTime: 'SYS:HH:MM:ss',
@@ -124,11 +31,18 @@ function buildCompileCmd (app) {
       })
     )
 
-    if (!await compile(fullPath, config, logger, compileOptions)) {
+    const compileOptions = {
+      cwd: fullPath,
+      logger,
+      clean: _args.includes('--clean'),
+      tsConfig: config.plugins?.typescript?.tsConfig,
+      flags: config.plugins?.typescript?.flags
+    }
+
+    if (!await compile(compileOptions)) {
       process.exit(1)
     }
   }
 }
 
-module.exports.compile = compile
 module.exports.buildCompileCmd = buildCompileCmd
