+++ packages/service/package.json
@@ -62,6 +62,7 @@
     "@platformatic/bus": "workspace:*",
     "@platformatic/client": "workspace:*",
     "@platformatic/config": "workspace:*",
+    "@platformatic/ts-compiler": "workspace:*",
     "@platformatic/generators": "workspace:*",
     "@platformatic/scalar-theme": "workspace:*",
     "@platformatic/telemetry": "workspace:*",
