+++ packages/ts-compiler/compile.js
@@ -1,0 +1,101 @@
+'use strict'
+
+const { resolve, join, dirname } = require('path')
+const pino = require('pino')
+const pretty = require('pino-pretty')
+const { isFileAccessible } = require('@platformatic/utils')
+const { readFile, rm } = require('fs/promises')
+
+async function getTSCExecutablePath (cwd) {
+  const typescriptPath = require.resolve('typescript')
+  const typescriptPathCWD = require.resolve('typescript', { paths: [process.cwd()] })
+
+  const tscLocalPath = join(typescriptPath, '..', '..', 'bin', 'tsc')
+  const tscGlobalPath = join(typescriptPathCWD, '..', '..', 'bin', 'tsc')
+
+  const [tscLocalExists, tscGlobalExists] = await Promise.all([
+    isFileAccessible(tscLocalPath),
+    isFileAccessible(tscGlobalPath)
+  ])
+
+  /* c8 ignore next 7 */
+  if (tscLocalExists) {
+    return tscLocalPath
+  }
+
+  if (tscGlobalExists) {
+    return tscGlobalPath
+  }
+}
+
+async function setup (options) {
+  if (!options.cwd) {
+    throw new Error('The cwd option is required.')
+  }
+
+  let logger = options.logger
+  if (!logger) {
+    logger = pino(
+      pretty({
+        translateTime: 'SYS:HH:MM:ss',
+        ignore: 'hostname,pid'
+      })
+    )
+  }
+
+  const { execa } = await import('execa')
+
+  const tscExecutablePath = await getTSCExecutablePath(options.cwd)
+  /* c8 ignore next 4 */
+  if (tscExecutablePath === undefined) {
+    const msg = 'The tsc executable was not found.'
+    logger.warn(msg)
+  }
+
+  const tsConfigPath = options.tsConfig || resolve(options.cwd, 'tsconfig.json')
+  const tsConfigExists = await isFileAccessible(tsConfigPath)
+
+  if (!tsConfigExists) {
+    const msg = 'No typescript configuration file was found, skipping compilation.'
+    logger.info(msg)
+  }
+
+  return { execa, logger, tscExecutablePath, tsConfigPath, tsConfigExists }
+}
+
+async function compile (options = {}) {
+  const { execa, logger, tscExecutablePath, tsConfigPath, tsConfigExists } = await setup(options)
+  /* c8 ignore next 3 */
+  if (!tscExecutablePath || !tsConfigExists) {
+    return false
+  }
+
+  try {
+    const tsFlags = options.flags || ['--project', tsConfigPath, '--rootDir', '.']
+    const env = {
+      ...process.env
+    }
+    delete env.NODE_V8_COVERAGE
+    // somehow c8 does not pick up these lines even if there is a specific test
+    /* c8 ignore start */
+    if (options && options.clean) {
+      // delete outdir directory
+      const tsConfigContents = JSON.parse(await readFile(tsConfigPath, 'utf8'))
+      const outDir = tsConfigContents.compilerOptions.outDir
+      if (outDir) {
+        const outDirFullPath = join(dirname(tsConfigPath), outDir)
+        logger.info(`Removing build directory ${outDirFullPath}`)
+        await rm(outDirFullPath, { recursive: true })
+      }
+    }
+    /* c8 ignore stop */
+    await execa(tscExecutablePath, tsFlags, { cwd: options.cwd, env })
+    logger.info('Typescript compilation completed successfully.')
+    return true
+  } catch (error) {
+    logger.error(error.message)
+    return false
+  }
+}
+
+module.exports.compile = compile
